[#2] Refactor: TravelProvider의 스코프 좁힘
기존 App에서 Main 컴포넌트로 좁히는 리팩토링
